Separate MindMapRow as a component file
